Release 1.1

+++ src/main/java/org/example/DevSync1/repository/TagRepository.java
@@ -1,18 +1,15 @@
 package org.example.DevSync1.repository;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.Persistence;
+import org.example.DevSync1.config.Config;
 import org.example.DevSync1.entity.Tag;
 
 import java.util.List;
 
 public class TagRepository {
 
-    private final EntityManagerFactory emf = Persistence.createEntityManagerFactory("default");
-
     public boolean save(Tag tag) {
-        EntityManager em = emf.createEntityManager();
+        EntityManager em = Config.getEntityManager();
         em.getTransaction().begin();
         em.persist(tag);
         em.getTransaction().commit();
@@ -21,7 +18,7 @@ public class TagRepository {
     }
 
     public boolean update(Tag tag) {
-        EntityManager em = emf.createEntityManager();
+        EntityManager em = Config.getEntityManager();
         em.getTransaction().begin();
         em.merge(tag);
         em.getTransaction().commit();
@@ -30,17 +27,19 @@ public class TagRepository {
     }
 
     public boolean delete(Long id) {
-        EntityManager em = emf.createEntityManager();
+        EntityManager em = Config.getEntityManager();
         em.getTransaction().begin();
         Tag tag = em.find(Tag.class, id);
-        em.remove(tag);
+        if (tag != null) {
+            em.remove(tag);
+        }
         em.getTransaction().commit();
         em.close();
         return true;
     }
 
     public List<Tag> findAll() {
-        EntityManager em = emf.createEntityManager();
+        EntityManager em = Config.getEntityManager();
         em.getTransaction().begin();
         List<Tag> tags = em.createQuery("SELECT c FROM Tag c ORDER BY c.id DESC", Tag.class).getResultList();
         em.getTransaction().commit();
@@ -49,12 +48,9 @@ public class TagRepository {
     }
 
     public Tag findById(Long id) {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
+        EntityManager em = Config.getEntityManager();
         Tag tag = em.find(Tag.class, id);
-        em.getTransaction().commit();
         em.close();
         return tag;
     }
-
 }
